changed migrations for heights

+++ db/migrate/20121008165654_change_data_type_for_height.rb
@@ -7,7 +7,7 @@ class ChangeDataTypeForHeight < ActiveRecord::Migration
 
   def down
   	change_table :people do |t|
-  		t.change :height, :integer
+  		t.change :height, :float
   	end
   end
 end
