+++ db/migrate/20121008202457_change_data_type_for_guess_height.rb
@@ -7,7 +7,7 @@ class ChangeDataTypeForGuessHeight < ActiveRecord::Migration
 
   def down
   	change_table :guesses do |t|
-  		t.change :height, :integer
+  		t.change :height, :float
   	end
   end
 end
